+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## 0.1.0-alpha.7 [Unreleased]
 ### Changed
   - Drop support of the php 5.6 Minimum version is php 7.0
-  
+
 ### Removed
+  - #40 Remove old methods: Fiv\Form\Element\Checkbox::unCheck, Fiv\Form\Element\Checkbox::check
   - #45 Remove method `Form::getData`
   
 ## 0.1.0-alpha.6 [2016-10-17]